Fix path failure preventing trace from opening
When doing some testing on an internal extension relying, I
noticed that we could no longer use the Open With TraceViewer
command on Windows due to an invalid path failing (worked on Linux).

I was able to trace the cause of this it to some code that uses
.path instead of .fsPath, which is intended for use with the
filesystem and scrubs out improper path values  (which is
what caused our error!)

This fixes a critical issue which prevents users from opening traces
in some cases.

Signed-off-by: Dylan Leclair <dleclair@blackberry.com>